requirements.txt
requests~=2.25.1
cryptography~=3.3.1
msal~=1.8.0
python-dateutil~=2.8.1
SQLAlchemy~=1.3.22
python-dotenv~=0.15.0
pandas~=1.2.1
pyodbc~=4.0.30
